+++ src/com/runnershigh/Util.java
@@ -17,11 +17,11 @@ public class Util {
 		return mInstance;
 	}
 	
-	public static float getPercentOfScreenWidth(int percent) {
+	public static float getPercentOfScreenWidth(float percent) {
 		float percentWidth=mScreenWidth/100*percent;
 		return percentWidth;
 	} 
-	public static float getPercentOfScreenHeight(int percent) {
+	public static float getPercentOfScreenHeight(float percent) {
 		float percentHeight=mScreenHeight/100*percent;
 		return percentHeight;
 	}
